initial attempt at making modules
