+++ rust/exo_rs/src/lib.rs
@@ -5,16 +5,11 @@
 //!
 
 mod allow_threading;
-mod ident;
-mod networking;
-mod pidfile;
+pub mod ident;
+pub mod networking;
+pub mod pidfile;
 
-use crate::ident::PyKeypair;
-use crate::networking::networking_submodule;
-use crate::pidfile::pidfile_submodule;
-use pyo3::prelude::PyModule;
-use pyo3::types::PyModuleMethods;
-use pyo3::{Bound, PyResult, pyclass, pymodule};
+use pyo3::{pyclass, pymodule};
 use pyo3_stub_gen::define_stub_info_gatherer;
 
 /// Namespace for all the constants used by this crate.
@@ -150,28 +145,31 @@ pub(crate) mod ext {
     }
 }
 
-/// A Python module implemented in Rust. The name of this function must match
-/// the `lib.name` setting in the `Cargo.toml`, else Python will not be able to
-/// import the module.
 #[pymodule(name = "exo_rs", gil_used = true)]
-fn main_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
-    // install logger
-    pyo3_log::init();
-    let mut builder = tokio::runtime::Builder::new_multi_thread();
-    builder.enable_all();
-    pyo3_async_runtimes::tokio::init(builder);
+mod py_exo_rs {
+    #[pymodule_export]
+    use super::ident::py_ident;
+    #[pymodule_export]
+    use super::networking::py_networking;
+    #[pymodule_export]
+    use super::pidfile::py_pidfile;
+    use pyo3::{
+        PyResult,
+        prelude::{Bound, PyModule},
+    };
 
-    // TODO: for now this is all NOT a submodule, but figure out how to make the submodule system
-    //       work with maturin, where the types generate correctly, in the right folder, without
-    //       too many importing issues...
-    m.add_class::<PyKeypair>()?;
-    networking_submodule(m)?;
-    pidfile_submodule(m)?;
+    #[pymodule_init]
+    fn init(_m: &Bound<'_, PyModule>) -> PyResult<()> {
+        // install logger (TODO: change to tracing)
+        pyo3_log::init();
 
-    // top-level constructs
-    // TODO: ...
+        // create pyo3_async_runtimes
+        let mut builder = tokio::runtime::Builder::new_multi_thread();
+        builder.enable_all();
+        pyo3_async_runtimes::tokio::init(builder);
 
-    Ok(())
+        Ok(())
+    }
 }
 
 define_stub_info_gatherer!(stub_info);
